+++ .storybook/preview.js
@@ -1,0 +1,14 @@
+import theme from "../src/styles/theme";
+
+export const parameters = {
+  actions: { argTypesRegex: "^on[A-Z].*" },
+  chakra: {
+    theme: theme,
+  },
+  controls: {
+    matchers: {
+      color: /(background|color)$/i,
+      date: /Date$/,
+    },
+  },
+};
