+++ src/components/SearchFactForm/SearchFactForm.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import {
   Button,
+  ButtonGroup,
   Flex,
   FormControl,
   Input,
@@ -8,37 +9,31 @@ import {
   useToast,
 } from "@chakra-ui/react";
 import axios from "axios";
-import { IFact } from "../SearchFact";
-import { IParamsFacts } from "../SearchListFacts";
+import { IFact } from "../SearchFact/SearchFact";
 
-export function SearchListFactsForm({
+export function SearchFactForm({
+  facts,
   setFacts,
   setLoading,
-  setParamsFacts,
-  setHasEndingPosts,
-  setCurrentPage,
 }: {
+  facts: IFact[] | null;
   setFacts: React.Dispatch<React.SetStateAction<IFact[] | null>>;
   setLoading: React.Dispatch<React.SetStateAction<boolean>>;
-  setParamsFacts: React.Dispatch<React.SetStateAction<IParamsFacts | null>>;
-  setHasEndingPosts: React.Dispatch<React.SetStateAction<boolean>>;
-  setCurrentPage: React.Dispatch<React.SetStateAction<number | null>>;
 }) {
-  const toast = useToast();
-  const [error, setError] = useState<boolean>(false);
   const [length, setLength] = useState<number | null>(null);
-  const [limit, setLimit] = useState<number | null>(null);
+  const [currentPage, setCurrentPage] = useState<number>(1);
+  const [hasEndingPosts, setHasEndingPosts] = useState<boolean>(false);
+  const [error, setError] = useState<boolean>(false);
+  const toast = useToast();
 
   const onChangeLength = (event: React.ChangeEvent<HTMLInputElement>) => {
     event.preventDefault();
-    setError(false);
-    setLength(parseInt(event.target.value, 10));
-  };
 
-  const onChangeLimit = (event: React.ChangeEvent<HTMLInputElement>) => {
-    event.preventDefault();
+    setCurrentPage(1);
     setError(false);
-    setLimit(parseInt(event.target.value, 10));
+    setHasEndingPosts(false);
+
+    setLength(parseInt(event.target.value, 10));
   };
 
   const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
@@ -52,18 +47,29 @@ export function SearchListFactsForm({
       .get("https://catfact.ninja/facts", {
         params: {
           max_length: length,
-          limit,
-          page: 1,
+          limit: 1,
+          page: currentPage,
         },
       })
       .then((res) => {
-        setParamsFacts({
-          max_length: length || null,
-          limit: limit || null,
-        });
-        setHasEndingPosts(false);
-        setCurrentPage(1);
-        setFacts(res.data.data);
+        if (currentPage === 1) {
+          setFacts(res.data.data);
+          setCurrentPage(currentPage + 1);
+          if (res.data.next_page_url === null) {
+            setHasEndingPosts(true);
+            setLoading(false);
+            return;
+          }
+          setLoading(false);
+          return;
+        }
+        setFacts([...(facts || []), ...res.data.data]);
+        setCurrentPage(currentPage + 1);
+        if (res.data.next_page_url === null) {
+          setHasEndingPosts(true);
+          setLoading(false);
+          return;
+        }
         setLoading(false);
       })
       .catch((err) => {
@@ -75,6 +81,14 @@ export function SearchListFactsForm({
         setLoading(false);
       });
   };
+
+  const onReset = () => {
+    setCurrentPage(1);
+    setLength(0);
+    setFacts(null);
+    setHasEndingPosts(false);
+  };
+
   return (
     <form onSubmit={(e) => handleSubmit(e)}>
       <FormControl id="length" isRequired>
@@ -85,26 +99,23 @@ export function SearchListFactsForm({
           onChange={(event) => onChangeLength(event)}
           value={length || ""}
         />
-      </FormControl>
-      <FormControl id="limit" marginTop="1vh" isRequired>
-        <Input
-          type="number"
-          name="limit"
-          placeholder="Quantidade de fatos"
-          onChange={(event) => onChangeLimit(event)}
-          value={limit || ""}
-        />
         {error && (
           <Text color="red">O tamanho precisa ser maior ou igual a 20</Text>
         )}
       </FormControl>
+
       <Flex justifyContent={["center", "center", "end"]}>
-        <Button marginTop="1vh" colorScheme="teal" type="submit">
-          Buscar
-        </Button>
+        <ButtonGroup marginTop="1vh" gap="1">
+          <Button type="button" onClick={onReset}>
+            Resetar
+          </Button>
+          <Button colorScheme="teal" type="submit" disabled={hasEndingPosts}>
+            Buscar
+          </Button>
+        </ButtonGroup>
       </Flex>
     </form>
   );
 }
 
-export default SearchListFactsForm;
+export default SearchFactForm;
